Correct nonce range

+++ applications/minotari_node/src/xmrig_proxy/error.rs
@@ -55,7 +55,7 @@ pub enum XmrigProxyError {
     #[error("Nonce out of assigned range: miner={0}, submitted={1}, range={2:?}")]
     // TODO: wire into handle_submit_block for nonce range verification (S2)
     #[allow(dead_code)]
-    NonceOutOfRange(String, u64, std::ops::Range<u32>),
+    NonceOutOfRange(String, u64, std::ops::Range<u64>),
 
     #[error("Max miners reached: {0}")]
     MaxMinersReached(usize),
